refactor(mysql): split statement policy modules
