+++ src/app/policy/sql/mysql_statement.rs
@@ -1,5 +1,13 @@
 use std::fmt;
 
+mod classifier;
+mod lexer;
+mod side_effect;
+mod target;
+mod transaction;
+
+use lexer::TokenKind;
+
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum MysqlStatementKind {
     Select,
@@ -44,783 +52,12 @@ impl fmt::Display for MysqlLexError {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum TokenKind {
-    Word(String),
-    Identifier(String),
-    StringLiteral,
-    Number,
-    Symbol(char),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct Token {
-    kind: TokenKind,
-    depth: usize,
-    text: String,
-}
-
 pub fn split_mysql_statements(sql: &str) -> Result<Vec<String>, MysqlLexError> {
-    let mut statements = Vec::new();
-    let mut start = 0;
-    let mut index = 0;
-    let mut quote = None;
-    let bytes = sql.as_bytes();
-    let mut depth = 0usize;
-
-    while index < bytes.len() {
-        let byte = bytes[index];
-        if let Some(delimiter) = quote {
-            if byte == b'\\' && delimiter != b'`' {
-                index += 2;
-                continue;
-            }
-            if byte == delimiter {
-                if bytes.get(index + 1) == Some(&delimiter) {
-                    index += 2;
-                } else {
-                    quote = None;
-                    index += 1;
-                }
-                continue;
-            }
-            index += 1;
-            continue;
-        }
-
-        if is_line_comment_start(bytes, index) {
-            index = skip_line_comment(bytes, index);
-            continue;
-        }
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            index = skip_block_comment(bytes, index)?;
-            continue;
-        }
-        if matches!(byte, b'\'' | b'"' | b'`') {
-            quote = Some(byte);
-            index += 1;
-            continue;
-        }
-        match byte {
-            b'(' => depth += 1,
-            b')' => {
-                depth = depth
-                    .checked_sub(1)
-                    .ok_or_else(|| MysqlLexError("unbalanced MySQL parentheses".to_string()))?;
-            }
-            b';' if depth == 0 => {
-                push_mysql_statement(&mut statements, &sql[start..index]);
-                start = index + 1;
-            }
-            _ => {}
-        }
-        index += 1;
-    }
-
-    if quote.is_some() {
-        return Err(MysqlLexError(
-            "unterminated MySQL quoted literal".to_string(),
-        ));
-    }
-    if depth != 0 {
-        return Err(MysqlLexError("unbalanced MySQL parentheses".to_string()));
-    }
-    push_mysql_statement(&mut statements, &sql[start..]);
-    Ok(statements)
-}
-
-fn push_mysql_statement(statements: &mut Vec<String>, fragment: &str) {
-    let trimmed = fragment.trim();
-    if !trimmed.is_empty() && !is_comment_only(trimmed) {
-        statements.push(trimmed.to_string());
-    }
-}
-
-fn is_line_comment_start(bytes: &[u8], index: usize) -> bool {
-    bytes[index] == b'#'
-        || (bytes.get(index..index + 2) == Some(b"--")
-            && bytes.get(index + 2).is_none_or(u8::is_ascii_whitespace))
-}
-
-fn skip_line_comment(bytes: &[u8], mut index: usize) -> usize {
-    while index < bytes.len() {
-        let byte = bytes[index];
-        index += 1;
-        if byte == b'\n' {
-            break;
-        }
-    }
-    index
-}
-
-fn skip_block_comment(bytes: &[u8], index: usize) -> Result<usize, MysqlLexError> {
-    let mut cursor = index + 2;
-    while cursor + 1 < bytes.len() {
-        if bytes[cursor] == b'*' && bytes[cursor + 1] == b'/' {
-            return Ok(cursor + 2);
-        }
-        cursor += 1;
-    }
-    Err(MysqlLexError(
-        "unterminated MySQL block comment".to_string(),
-    ))
-}
-
-fn is_comment_only(sql: &str) -> bool {
-    let bytes = sql.as_bytes();
-    let mut index = 0;
-    while index < bytes.len() {
-        if bytes[index].is_ascii_whitespace() {
-            index += 1;
-        } else if is_line_comment_start(bytes, index) {
-            index = skip_line_comment(bytes, index);
-        } else if bytes.get(index..index + 2) == Some(b"/*") {
-            if bytes.get(index + 2) == Some(&b'!') {
-                return false;
-            }
-            index = match skip_block_comment(bytes, index) {
-                Ok(next) => next,
-                Err(_) => return false,
-            };
-        } else {
-            return false;
-        }
-    }
-    true
-}
-
-fn lex_mysql_statement(sql: &str) -> Result<Vec<Token>, MysqlLexError> {
-    let bytes = sql.as_bytes();
-    let mut tokens = Vec::new();
-    let mut index = 0;
-    let mut depth = 0usize;
-    let mut leading_executable_version_comment = false;
-
-    while index < bytes.len() {
-        if bytes[index].is_ascii_whitespace() {
-            index += 1;
-            continue;
-        }
-        if is_line_comment_start(bytes, index) {
-            index = skip_line_comment(bytes, index);
-            continue;
-        }
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            let end = skip_block_comment(bytes, index)?;
-            let body = &sql[index + 2..end - 2];
-            if let Some(body) = body.strip_prefix('!') {
-                let body = body.trim_start();
-                let version_len = body.bytes().take_while(u8::is_ascii_digit).count();
-                if version_len == 0 {
-                    return Err(MysqlLexError(
-                        "MySQL version comment has no verifiable version".to_string(),
-                    ));
-                }
-                let content = body[version_len..].trim_start();
-                if content.is_empty() {
-                    return Err(MysqlLexError(
-                        "MySQL version comment has no executable statement".to_string(),
-                    ));
-                }
-                let inner = lex_mysql_statement(content)?;
-                let executable_statement = inner.first().is_some_and(|token| {
-                    matches!(&token.kind, TokenKind::Word(word) if is_mysql_statement_keyword(word))
-                });
-                let contains_statement_separator = inner
-                    .iter()
-                    .any(|token| matches!(token.kind, TokenKind::Symbol(';')));
-                if tokens.is_empty() && executable_statement && !contains_statement_separator {
-                    tokens.extend(inner);
-                    leading_executable_version_comment = true;
-                } else if !inner.is_empty() {
-                    tokens.push(Token {
-                        kind: TokenKind::Word("UNSUPPORTED_VERSION_COMMENT".to_string()),
-                        depth,
-                        text: "UNSUPPORTED_VERSION_COMMENT".to_string(),
-                    });
-                }
-            }
-            index = end;
-            continue;
-        }
-
-        if leading_executable_version_comment {
-            tokens.push(Token {
-                kind: TokenKind::Word("UNSUPPORTED_VERSION_COMMENT".to_string()),
-                depth,
-                text: "UNSUPPORTED_VERSION_COMMENT".to_string(),
-            });
-            leading_executable_version_comment = false;
-        }
-
-        let byte = bytes[index];
-        if matches!(byte, b'\'' | b'"' | b'`') {
-            let end = skip_quoted(bytes, index, byte)?;
-            let text = &sql[index + 1..end - 1];
-            let kind = if byte == b'`' {
-                TokenKind::Identifier(text.replace("``", "`"))
-            } else {
-                TokenKind::StringLiteral
-            };
-            tokens.push(Token {
-                kind,
-                depth,
-                text: text.to_string(),
-            });
-            index = end;
-            continue;
-        }
-        if byte.is_ascii_alphabetic() || byte == b'_' || byte == b'$' {
-            let start = index;
-            index += 1;
-            while index < bytes.len()
-                && (bytes[index].is_ascii_alphanumeric() || matches!(bytes[index], b'_' | b'$'))
-            {
-                index += 1;
-            }
-            let text = sql[start..index].to_string();
-            tokens.push(Token {
-                kind: TokenKind::Word(text.to_ascii_uppercase()),
-                depth,
-                text,
-            });
-            continue;
-        }
-        if byte.is_ascii_digit() {
-            let start = index;
-            index += 1;
-            while index < bytes.len()
-                && (bytes[index].is_ascii_alphanumeric() || matches!(bytes[index], b'.' | b'_'))
-            {
-                index += 1;
-            }
-            tokens.push(Token {
-                kind: TokenKind::Number,
-                depth,
-                text: sql[start..index].to_string(),
-            });
-            continue;
-        }
-        let kind = TokenKind::Symbol(byte as char);
-        tokens.push(Token {
-            kind,
-            depth,
-            text: (byte as char).to_string(),
-        });
-        match byte {
-            b'(' => depth += 1,
-            b')' => {
-                depth = depth
-                    .checked_sub(1)
-                    .ok_or_else(|| MysqlLexError("unbalanced MySQL parentheses".to_string()))?;
-            }
-            _ => {}
-        }
-        index += 1;
-    }
-    Ok(tokens)
-}
-
-fn is_mysql_statement_keyword(word: &str) -> bool {
-    matches!(
-        word,
-        "SELECT"
-            | "TABLE"
-            | "SHOW"
-            | "DESCRIBE"
-            | "DESC"
-            | "WITH"
-            | "INSERT"
-            | "UPDATE"
-            | "DELETE"
-            | "CREATE"
-            | "ALTER"
-            | "DROP"
-            | "TRUNCATE"
-            | "BEGIN"
-            | "START"
-            | "COMMIT"
-            | "ROLLBACK"
-            | "SAVEPOINT"
-            | "RELEASE"
-            | "USE"
-            | "SET"
-            | "LOCK"
-            | "UNLOCK"
-            | "REPLACE"
-            | "CALL"
-            | "DO"
-            | "HANDLER"
-            | "LOAD"
-            | "PREPARE"
-            | "EXECUTE"
-            | "DEALLOCATE"
-            | "XA"
-    )
-}
-
-fn skip_quoted(bytes: &[u8], index: usize, quote: u8) -> Result<usize, MysqlLexError> {
-    let mut cursor = index + 1;
-    while cursor < bytes.len() {
-        if bytes[cursor] == b'\\' && quote != b'`' {
-            cursor += 2;
-            continue;
-        }
-        if bytes[cursor] == quote {
-            if bytes.get(cursor + 1) == Some(&quote) {
-                cursor += 2;
-            } else {
-                return Ok(cursor + 1);
-            }
-        } else {
-            cursor += 1;
-        }
-    }
-    Err(MysqlLexError(
-        "unterminated MySQL quoted literal".to_string(),
-    ))
-}
-
-fn word(tokens: &[Token], index: usize) -> Option<&str> {
-    match tokens.get(index)?.kind {
-        TokenKind::Word(ref word) => Some(word),
-        _ => None,
-    }
-}
-
-fn top_level_word(tokens: &[Token], word_value: &str) -> bool {
-    tokens.iter().any(|token| {
-        token.depth == 0 && matches!(&token.kind, TokenKind::Word(word) if word == word_value)
-    })
-}
-
-fn identifier_at(tokens: &[Token], mut index: usize) -> Option<(String, Option<String>, usize)> {
-    while matches!(
-        tokens.get(index).map(|token| &token.kind),
-        Some(TokenKind::Symbol(','))
-    ) {
-        index += 1;
-    }
-    let first_token = tokens.get(index)?;
-    let first = match &first_token.kind {
-        TokenKind::Word(_) => first_token.text.clone(),
-        TokenKind::Identifier(identifier) => identifier.clone(),
-        _ => return None,
-    };
-    if matches!(
-        tokens.get(index + 1).map(|token| &token.kind),
-        Some(TokenKind::Symbol('.'))
-    ) {
-        let second_token = tokens.get(index + 2)?;
-        let second = match &second_token.kind {
-            TokenKind::Word(_) => second_token.text.clone(),
-            TokenKind::Identifier(identifier) => identifier.clone(),
-            _ => return None,
-        };
-        return Some((second, Some(first), index + 3));
-    }
-    Some((first, None, index + 1))
-}
-
-fn skip_mysql_modifiers(tokens: &[Token], mut index: usize, modifiers: &[&str]) -> usize {
-    while word(tokens, index).is_some_and(|value| modifiers.contains(&value)) {
-        index += 1;
-    }
-    index
-}
-
-fn find_word(tokens: &[Token], value: &str, start: usize) -> Option<usize> {
-    tokens
-        .iter()
-        .enumerate()
-        .skip(start)
-        .find_map(|(index, token)| {
-            (token.depth == 0 && matches!(&token.kind, TokenKind::Word(word) if word == value))
-                .then_some(index)
-        })
-}
-
-fn cte_body_start(tokens: &[Token]) -> Option<usize> {
-    if word(tokens, 0) != Some("WITH") {
-        return None;
-    }
-    let mut index = 1;
-    if word(tokens, index) == Some("RECURSIVE") {
-        index += 1;
-    }
-    loop {
-        if !matches!(
-            tokens.get(index).map(|token| &token.kind),
-            Some(TokenKind::Word(_) | TokenKind::Identifier(_))
-        ) {
-            return None;
-        }
-        index += 1;
-        if matches!(
-            tokens.get(index).map(|token| &token.kind),
-            Some(TokenKind::Symbol('('))
-        ) {
-            index = skip_parenthesized_tokens(tokens, index)?;
-        }
-        if word(tokens, index) != Some("AS") {
-            return None;
-        }
-        index += 1;
-        if !matches!(
-            tokens.get(index).map(|token| &token.kind),
-            Some(TokenKind::Symbol('('))
-        ) {
-            return None;
-        }
-        index = skip_parenthesized_tokens(tokens, index)?;
-        if matches!(
-            tokens.get(index).map(|token| &token.kind),
-            Some(TokenKind::Symbol(','))
-        ) {
-            index += 1;
-            continue;
-        }
-        return Some(index);
-    }
-}
-
-fn skip_parenthesized_tokens(tokens: &[Token], index: usize) -> Option<usize> {
-    if !matches!(
-        tokens.get(index).map(|token| &token.kind),
-        Some(TokenKind::Symbol('('))
-    ) {
-        return None;
-    }
-    let mut depth = 0usize;
-    for (cursor, token) in tokens.iter().enumerate().skip(index) {
-        match token.kind {
-            TokenKind::Symbol('(') => depth += 1,
-            TokenKind::Symbol(')') => {
-                depth = depth.checked_sub(1)?;
-                if depth == 0 {
-                    return Some(cursor + 1);
-                }
-            }
-            _ => {}
-        }
-    }
-    None
-}
-
-fn drop_target_after(
-    tokens: &[Token],
-    index: usize,
-) -> Result<(Option<String>, Option<String>), MysqlLexError> {
-    let (target, database, next) = identifier_at(tokens, index)
-        .ok_or_else(|| MysqlLexError("MySQL statement target is ambiguous".to_string()))?;
-    if tokens[next..]
-        .iter()
-        .any(|token| token.depth == 0 && matches!(token.kind, TokenKind::Symbol(',')))
-    {
-        return Err(MysqlLexError(
-            "MySQL DROP statements must have one target".to_string(),
-        ));
-    }
-    Ok((Some(target), database))
-}
-
-fn effective_start(tokens: &[Token]) -> usize {
-    cte_body_start(tokens).unwrap_or(0)
-}
-
-type MysqlClassification = (MysqlStatementKind, Option<String>, Option<String>);
-
-fn target_after(
-    tokens: &[Token],
-    index: usize,
-) -> Result<(Option<String>, Option<String>), MysqlLexError> {
-    identifier_at(tokens, index)
-        .map(|(target, database, _)| (Some(target), database))
-        .ok_or_else(|| MysqlLexError("MySQL statement target is ambiguous".to_string()))
-}
-
-fn classify_mysql_crud_statement(
-    tokens: &[Token],
-    start: usize,
-    first: &str,
-) -> Result<MysqlClassification, MysqlLexError> {
-    match first {
-        "SELECT" => Ok((MysqlStatementKind::Select, None, None)),
-        "INSERT" | "REPLACE" => {
-            let index = if first == "INSERT" {
-                skip_mysql_modifiers(
-                    tokens,
-                    start + 1,
-                    &["LOW_PRIORITY", "DELAYED", "HIGH_PRIORITY", "IGNORE"],
-                )
-            } else {
-                skip_mysql_modifiers(tokens, start + 1, &["LOW_PRIORITY", "DELAYED"])
-            };
-            let target_index = if word(tokens, index) == Some("INTO") {
-                index + 1
-            } else {
-                index
-            };
-            let (target, database) = target_after(tokens, target_index)?;
-            let kind = if first == "REPLACE" {
-                MysqlStatementKind::Replace
-            } else {
-                MysqlStatementKind::Insert
-            };
-            Ok((kind, target, database))
-        }
-        "UPDATE" => {
-            let has_where = top_level_word(&tokens[start..], "WHERE");
-            let target_index = skip_mysql_modifiers(tokens, start + 1, &["LOW_PRIORITY", "IGNORE"]);
-            let (target, database) = target_after(tokens, target_index)?;
-            Ok((MysqlStatementKind::Update { has_where }, target, database))
-        }
-        "DELETE" => {
-            let has_where = top_level_word(&tokens[start..], "WHERE");
-            let index =
-                skip_mysql_modifiers(tokens, start + 1, &["LOW_PRIORITY", "QUICK", "IGNORE"]);
-            let target_index = if word(tokens, index) == Some("FROM") {
-                index + 1
-            } else {
-                index
-            };
-            let (target, database) = target_after(tokens, target_index)?;
-            Ok((MysqlStatementKind::Delete { has_where }, target, database))
-        }
-        _ => unreachable!("not a MySQL CRUD statement: {first}"),
-    }
-}
-
-fn classify_mysql_ddl_statement(
-    tokens: &[Token],
-    start: usize,
-    first: &str,
-) -> Result<MysqlClassification, MysqlLexError> {
-    match first {
-        "CREATE" => {
-            let mut index = start + 1;
-            let temporary = word(tokens, index) == Some("TEMPORARY");
-            if temporary {
-                index += 1;
-            }
-            match word(tokens, index) {
-                Some("TABLE") => {
-                    index += 1;
-                    if word(tokens, index) == Some("IF") {
-                        index += 3;
-                    }
-                    let (target, database) = target_after(tokens, index)?;
-                    Ok((
-                        MysqlStatementKind::CreateTable { temporary },
-                        target,
-                        database,
-                    ))
-                }
-                Some("VIEW") => {
-                    let (target, database) = target_after(tokens, index + 1)?;
-                    Ok((MysqlStatementKind::CreateView, target, database))
-                }
-                Some("UNIQUE") if word(tokens, index + 1) == Some("INDEX") => {
-                    let on = find_word(tokens, "ON", index + 2).ok_or_else(|| {
-                        MysqlLexError("CREATE INDEX target is ambiguous".to_string())
-                    })?;
-                    let (_, database) = target_after(tokens, on + 1)?;
-                    let (index_name, _, _) = identifier_at(tokens, index + 2).ok_or_else(|| {
-                        MysqlLexError("CREATE INDEX name is ambiguous".to_string())
-                    })?;
-                    Ok((MysqlStatementKind::CreateIndex, Some(index_name), database))
-                }
-                Some("INDEX" | "KEY") => {
-                    let on = find_word(tokens, "ON", index + 1).ok_or_else(|| {
-                        MysqlLexError("CREATE INDEX target is ambiguous".to_string())
-                    })?;
-                    let (_, database) = target_after(tokens, on + 1)?;
-                    let (index_name, _, _) = identifier_at(tokens, index + 1).ok_or_else(|| {
-                        MysqlLexError("CREATE INDEX name is ambiguous".to_string())
-                    })?;
-                    Ok((MysqlStatementKind::CreateIndex, Some(index_name), database))
-                }
-                _ => Err(MysqlLexError(
-                    "unsupported MySQL CREATE statement".to_string(),
-                )),
-            }
-        }
-        "ALTER" => {
-            if word(tokens, start + 1) != Some("TABLE") {
-                return Err(MysqlLexError(
-                    "unsupported MySQL ALTER statement".to_string(),
-                ));
-            }
-            let (target, database) = target_after(tokens, start + 2)?;
-            Ok((MysqlStatementKind::AlterTable, target, database))
-        }
-        "DROP" => {
-            let mut index = start + 1;
-            let temporary = word(tokens, index) == Some("TEMPORARY");
-            if temporary {
-                index += 1;
-            }
-            match word(tokens, index) {
-                Some("TABLE") => {
-                    let target_index = if word(tokens, index + 1) == Some("IF") {
-                        index + 3
-                    } else {
-                        index + 1
-                    };
-                    let (target, database) = drop_target_after(tokens, target_index)?;
-                    Ok((
-                        MysqlStatementKind::DropTable { temporary },
-                        target,
-                        database,
-                    ))
-                }
-                Some("VIEW") => {
-                    let target_index = if word(tokens, index + 1) == Some("IF") {
-                        index + 3
-                    } else {
-                        index + 1
-                    };
-                    let (target, database) = drop_target_after(tokens, target_index)?;
-                    Ok((MysqlStatementKind::DropView, target, database))
-                }
-                Some("INDEX" | "KEY") => {
-                    let on = find_word(tokens, "ON", index + 1).ok_or_else(|| {
-                        MysqlLexError("DROP INDEX target is ambiguous".to_string())
-                    })?;
-                    let (_, database) = target_after(tokens, on + 1)?;
-                    let index_name_index = if word(tokens, index + 1) == Some("IF") {
-                        index + 3
-                    } else {
-                        index + 1
-                    };
-                    let (index_name, _, _) = identifier_at(tokens, index_name_index)
-                        .ok_or_else(|| MysqlLexError("DROP INDEX name is ambiguous".to_string()))?;
-                    Ok((MysqlStatementKind::DropIndex, Some(index_name), database))
-                }
-                _ => Err(MysqlLexError(
-                    "unsupported MySQL DROP statement".to_string(),
-                )),
-            }
-        }
-        "TRUNCATE" => {
-            let target_index = if word(tokens, start + 1) == Some("TABLE") {
-                start + 2
-            } else {
-                start + 1
-            };
-            let (target, database) = target_after(tokens, target_index)?;
-            Ok((MysqlStatementKind::TruncateTable, target, database))
-        }
-        _ => unreachable!("not a MySQL DDL statement: {first}"),
-    }
-}
-
-fn classify_mysql_transaction_statement(
-    tokens: &[Token],
-    start: usize,
-    first: &str,
-) -> Result<MysqlClassification, MysqlLexError> {
-    match first {
-        "BEGIN" => {
-            if tokens.len() != start + 1 {
-                return Err(MysqlLexError(
-                    "BEGIN modifiers are not supported".to_string(),
-                ));
-            }
-            Ok((MysqlStatementKind::Begin, None, None))
-        }
-        "START" if word(tokens, start + 1) == Some("TRANSACTION") => {
-            if tokens.len() != start + 2 {
-                return Err(MysqlLexError(
-                    "START TRANSACTION modifiers are not supported".to_string(),
-                ));
-            }
-            Ok((MysqlStatementKind::StartTransaction, None, None))
-        }
-        "COMMIT" => {
-            if tokens.len() != start + 1 {
-                return Err(MysqlLexError(
-                    "COMMIT modifiers are not supported".to_string(),
-                ));
-            }
-            Ok((MysqlStatementKind::Commit, None, None))
-        }
-        "ROLLBACK" => {
-            if word(tokens, start + 1).is_none() && tokens.len() == start + 1 {
-                Ok((MysqlStatementKind::Rollback, None, None))
-            } else if word(tokens, start + 1) == Some("TO") {
-                let name_index = if word(tokens, start + 2) == Some("SAVEPOINT") {
-                    start + 3
-                } else {
-                    start + 2
-                };
-                if tokens.len() != name_index + 1 {
-                    return Err(MysqlLexError(
-                        "ROLLBACK modifiers are not supported".to_string(),
-                    ));
-                }
-                let (name, database, _) = identifier_at(tokens, name_index).ok_or_else(|| {
-                    MysqlLexError("ROLLBACK SAVEPOINT name is ambiguous".to_string())
-                })?;
-                Ok((
-                    MysqlStatementKind::RollbackToSavepoint,
-                    Some(name),
-                    database,
-                ))
-            } else {
-                Err(MysqlLexError(
-                    "ROLLBACK modifiers are not supported".to_string(),
-                ))
-            }
-        }
-        "SAVEPOINT" if tokens.len() == start + 2 => {
-            let (name, database, _) = identifier_at(tokens, start + 1)
-                .ok_or_else(|| MysqlLexError("SAVEPOINT name is ambiguous".to_string()))?;
-            Ok((MysqlStatementKind::Savepoint, Some(name), database))
-        }
-        "RELEASE" if word(tokens, start + 1) == Some("SAVEPOINT") && tokens.len() == start + 3 => {
-            let (name, database, _) = identifier_at(tokens, start + 2)
-                .ok_or_else(|| MysqlLexError("RELEASE SAVEPOINT name is ambiguous".to_string()))?;
-            Ok((MysqlStatementKind::ReleaseSavepoint, Some(name), database))
-        }
-        _ => Err(MysqlLexError(format!(
-            "unsupported MySQL statement: {first}"
-        ))),
-    }
-}
-
-fn classify_mysql_utility_statement(first: &str) -> MysqlClassification {
-    match first {
-        "TABLE" => (MysqlStatementKind::Table, None, None),
-        "SHOW" => (MysqlStatementKind::Show, None, None),
-        "DESCRIBE" | "DESC" => (MysqlStatementKind::Describe, None, None),
-        _ => unreachable!("not a MySQL utility statement: {first}"),
-    }
-}
-
-fn kind_and_target(tokens: &[Token]) -> Result<MysqlClassification, MysqlLexError> {
-    let start = effective_start(tokens);
-    let first =
-        word(tokens, start).ok_or_else(|| MysqlLexError("unknown MySQL statement".to_string()))?;
-    match first {
-        "SELECT" | "INSERT" | "REPLACE" | "UPDATE" | "DELETE" => {
-            classify_mysql_crud_statement(tokens, start, first)
-        }
-        "CREATE" | "ALTER" | "DROP" | "TRUNCATE" => {
-            classify_mysql_ddl_statement(tokens, start, first)
-        }
-        "BEGIN" | "START" | "COMMIT" | "ROLLBACK" | "SAVEPOINT" | "RELEASE" => {
-            classify_mysql_transaction_statement(tokens, start, first)
-        }
-        "TABLE" | "SHOW" | "DESCRIBE" | "DESC" => Ok(classify_mysql_utility_statement(first)),
-        _ => Err(MysqlLexError(format!(
-            "unsupported MySQL statement: {first}"
-        ))),
-    }
+    lexer::split_mysql_statements(sql)
 }
 
 pub fn classify_mysql_statement(sql: &str) -> Result<MysqlStatement, MysqlLexError> {
-    let tokens = lex_mysql_statement(sql)?;
+    let tokens = lexer::lex_mysql_statement(sql)?;
     if tokens.is_empty() {
         return Err(MysqlLexError("empty MySQL statement".to_string()));
     }
@@ -834,13 +71,36 @@ pub fn classify_mysql_statement(sql: &str) -> Result<MysqlStatement, MysqlLexErr
             "executable MySQL version comment contains another statement".to_string(),
         ));
     }
-    let (kind, target, target_database) = kind_and_target(&tokens)?;
+    let (kind, target, target_database) = classifier::kind_and_target(&tokens)?;
     Ok(MysqlStatement {
         sql: sql.to_string(),
         kind,
         target,
         target_database,
     })
+}
+
+pub fn has_top_level_into_clause(sql: &str) -> Result<bool, MysqlLexError> {
+    side_effect::has_top_level_into_clause(sql)
+}
+
+pub fn has_mysql_read_only_side_effect(sql: &str) -> Result<bool, MysqlLexError> {
+    side_effect::has_mysql_read_only_side_effect(sql)
+}
+
+pub fn has_mysql_version_comment(sql: &str) -> Result<bool, MysqlLexError> {
+    side_effect::has_mysql_version_comment(sql)
+}
+
+pub fn target_is_selected_database(
+    statement: &MysqlStatement,
+    selected_database: Option<&str>,
+) -> bool {
+    target::target_is_selected_database(statement, selected_database)
+}
+
+pub fn statement_contains_unsupported_mysql_control(sql: &str) -> bool {
+    side_effect::statement_contains_unsupported_mysql_control(sql)
 }
 
 pub fn mysql_explain_rejection_message(sql: &str) -> Option<&'static str> {
@@ -953,224 +213,6 @@ fn consume_required_ascii_whitespace(text: &str, index: &mut usize) -> Option<()
     let start = *index;
     skip_ascii_whitespace(text, index);
     (*index > start).then_some(())
-}
-
-pub fn has_top_level_into_clause(sql: &str) -> Result<bool, MysqlLexError> {
-    let tokens = lex_mysql_statement(sql)?;
-    Ok(tokens.iter().any(|token| {
-        token.depth == 0 && matches!(&token.kind, TokenKind::Word(word) if word == "INTO")
-    }))
-}
-
-pub fn has_mysql_read_only_side_effect(sql: &str) -> Result<bool, MysqlLexError> {
-    if has_mysql_version_comment(sql)? {
-        return Ok(true);
-    }
-
-    let tokens = lex_mysql_statement(sql)?;
-    for (index, token) in tokens.iter().enumerate() {
-        if matches!(&token.kind, TokenKind::Word(word) if word == "INTO") {
-            return Ok(true);
-        }
-        if matches!(&token.kind, TokenKind::Word(word) if matches!(
-            word.as_str(),
-            "GET_LOCK" | "RELEASE_LOCK" | "RELEASE_ALL_LOCKS"
-        )) {
-            return Ok(true);
-        }
-        if matches!(&token.kind, TokenKind::Symbol(':'))
-            && matches!(
-                tokens.get(index + 1).map(|token| &token.kind),
-                Some(TokenKind::Symbol('='))
-            )
-        {
-            return Ok(true);
-        }
-        if word(&tokens, index) == Some("FOR")
-            && matches!(word(&tokens, index + 1), Some("UPDATE" | "SHARE"))
-        {
-            return Ok(true);
-        }
-        if word(&tokens, index) == Some("LOCK")
-            && word(&tokens, index + 1) == Some("IN")
-            && word(&tokens, index + 2) == Some("SHARE")
-            && word(&tokens, index + 3) == Some("MODE")
-        {
-            return Ok(true);
-        }
-    }
-    Ok(false)
-}
-
-pub fn has_mysql_version_comment(sql: &str) -> Result<bool, MysqlLexError> {
-    let bytes = sql.as_bytes();
-    let mut index = 0;
-    let mut quote = None;
-
-    while index < bytes.len() {
-        if let Some(delimiter) = quote {
-            if bytes[index] == b'\\' && delimiter != b'`' {
-                index = (index + 2).min(bytes.len());
-            } else if bytes[index] == delimiter {
-                if bytes.get(index + 1) == Some(&delimiter) {
-                    index += 2;
-                } else {
-                    quote = None;
-                    index += 1;
-                }
-            } else {
-                index += 1;
-            }
-            continue;
-        }
-        if is_line_comment_start(bytes, index) {
-            index = skip_line_comment(bytes, index);
-            continue;
-        }
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            if bytes.get(index + 2) == Some(&b'!') {
-                return Ok(true);
-            }
-            index = skip_block_comment(bytes, index)?;
-            continue;
-        }
-        if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            quote = Some(bytes[index]);
-        }
-        index += 1;
-    }
-
-    if quote.is_some() {
-        return Err(MysqlLexError(
-            "unterminated MySQL quoted literal".to_string(),
-        ));
-    }
-    Ok(false)
-}
-
-pub fn target_is_selected_database(
-    statement: &MysqlStatement,
-    selected_database: Option<&str>,
-) -> bool {
-    match (statement.target_database.as_deref(), selected_database) {
-        (Some(target_database), Some(selected)) => target_database.eq_ignore_ascii_case(selected),
-        (None, Some(_)) => true,
-        (Some(_) | None, None) => false,
-    }
-}
-
-pub fn statement_contains_unsupported_mysql_control(sql: &str) -> bool {
-    let lower = sql.trim_start().to_ascii_lowercase();
-    [
-        "use ",
-        "use\n",
-        "set ",
-        "set\n",
-        "lock tables",
-        "unlock tables",
-    ]
-    .iter()
-    .any(|prefix| lower.starts_with(prefix))
-        || contains_mysql_client_command(sql)
-}
-
-fn contains_mysql_client_command(sql: &str) -> bool {
-    let bytes = sql.as_bytes();
-    let mut index = 0;
-    let mut line_start = true;
-    let mut quote = None;
-    let mut block_comment = false;
-
-    while index < bytes.len() {
-        if let Some(delimiter) = quote {
-            if bytes[index] == b'\\' && delimiter != b'`' {
-                index = (index + 2).min(bytes.len());
-            } else if bytes[index] == delimiter {
-                if bytes.get(index + 1) == Some(&delimiter) {
-                    index += 2;
-                } else {
-                    quote = None;
-                    index += 1;
-                }
-            } else {
-                index += 1;
-            }
-            continue;
-        }
-        if block_comment {
-            if bytes.get(index..index + 2) == Some(b"*/") {
-                block_comment = false;
-                index += 2;
-            } else {
-                line_start = bytes[index] == b'\n' || line_start;
-                index += 1;
-            }
-            continue;
-        }
-        if bytes[index] == b'\\' {
-            return true;
-        }
-        if line_start {
-            if matches!(bytes[index], b' ' | b'\t' | b'\r') {
-                index += 1;
-                continue;
-            }
-            if bytes[index] == b'\n' {
-                index += 1;
-                continue;
-            }
-            if bytes[index] == b'\\' {
-                return true;
-            }
-            if bytes.get(index..index + 2) == Some(b"/*") {
-                block_comment = true;
-                index += 2;
-                continue;
-            }
-            if is_line_comment_start(bytes, index) {
-                index = skip_line_comment(bytes, index);
-                continue;
-            }
-            if bytes[index].is_ascii_alphabetic() {
-                let start = index;
-                index += 1;
-                while index < bytes.len()
-                    && (bytes[index].is_ascii_alphanumeric() || bytes[index] == b'_')
-                {
-                    index += 1;
-                }
-                let word = &sql[start..index];
-                if matches!(
-                    word.to_ascii_lowercase().as_str(),
-                    "delimiter" | "charset" | "source" | "system"
-                ) && (index == bytes.len() || bytes[index].is_ascii_whitespace())
-                {
-                    return true;
-                }
-                line_start = false;
-                continue;
-            }
-            line_start = false;
-        }
-
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            block_comment = true;
-            index += 2;
-        } else if is_line_comment_start(bytes, index) {
-            index = skip_line_comment(bytes, index);
-            line_start = true;
-        } else if bytes[index] == b'\n' {
-            line_start = true;
-            index += 1;
-        } else if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            quote = Some(bytes[index]);
-            line_start = false;
-            index += 1;
-        } else {
-            index += 1;
-        }
-    }
-    false
 }
 
 #[cfg(test)]
